feat: create seperate component for order preview
