+++ apps/vardast-client/src/app/(client)/(home)/components/PublicPreOrderCard.tsx
@@ -1,9 +1,7 @@
-import { digitsEnToFa } from "@persian-tools/persian-tools"
-import { DetailsWithTitle } from "@vardast/component/desktop/DetailsWithTitle"
-import DynamicHeroIcon from "@vardast/component/DynamicHeroIcon"
-import { PublicPreOrderDto } from "@vardast/graphql/generated"
-import { mergeClasses } from "@vardast/tailwind-config/mergeClasses"
+import { PreOrderDto, PublicPreOrderDto } from "@vardast/graphql/generated"
 import useTranslation from "next-translate/useTranslation"
+
+import OrderPreviewCard from "@/app/(client)/(home)/components/OrderPreviewCard"
 
 type Props = { data: PublicPreOrderDto }
 
@@ -44,82 +42,7 @@ function PublicPreOrderCard({ data }: Props) {
       <div className="flex flex-col divide-y border-l px-5">
         <span className="pb-5 text-lg font-semibold">{data.categoryName}</span>
         {data.orders.map((order, index) => (
-          <div className="flex flex-col gap-3 py-5">
-            <div
-              key={order.id}
-              className="flex w-full items-center justify-between"
-            >
-              <div className="flex items-center gap-2">
-                <DynamicHeroIcon
-                  icon="ClipboardDocumentListIcon"
-                  className={mergeClasses(
-                    "icon h-7 w-7 transform rounded-md bg-primary-600 p-1 text-alpha-white transition-all"
-                  )}
-                  solid={false}
-                />
-
-                <span className="font-semibold">
-                  سفارش شماره ({digitsEnToFa(order.uuid)})
-                </span>
-              </div>
-              <span className="tag tag-info text-xs">
-                {digitsEnToFa(
-                  new Date(order.request_date).toLocaleDateString("fa-IR", {
-                    year: "numeric",
-                    month: "2-digit",
-                    day: "2-digit"
-                  })
-                )}
-              </span>
-            </div>
-            <ol className="gap-2">
-              <li>
-                {/* <DetailsWithTitle title={t("common:destination")} /> */}
-              </li>
-              <li>
-                <DetailsWithTitle
-                  title={t("common:needed-time")}
-                  text={digitsEnToFa(
-                    new Date(order.need_date).toLocaleDateString("fa-IR", {
-                      year: "numeric",
-                      month: "2-digit",
-                      day: "2-digit",
-                      hour: "2-digit",
-                      minute: "2-digit"
-                    })
-                  )}
-                />
-              </li>
-              <li>
-                <DetailsWithTitle
-                  title={t("common:bid-start-time")}
-                  text={digitsEnToFa(
-                    new Date(order.bid_start).toLocaleDateString("fa-IR", {
-                      year: "numeric",
-                      month: "2-digit",
-                      day: "2-digit",
-                      hour: "2-digit",
-                      minute: "2-digit"
-                    })
-                  )}
-                />
-              </li>
-              <li>
-                <DetailsWithTitle
-                  title={t("common:bid-end-time")}
-                  text={digitsEnToFa(
-                    new Date(order.bid_end).toLocaleDateString("fa-IR", {
-                      year: "numeric",
-                      month: "2-digit",
-                      day: "2-digit",
-                      hour: "2-digit",
-                      minute: "2-digit"
-                    })
-                  )}
-                />
-              </li>
-            </ol>
-          </div>
+          <OrderPreviewCard key={order.id} order={order as PreOrderDto} />
         ))}
       </div>
     )
